TransceiverUHD: Correct receive RRC pulse shaping filter bandwidth
Existing implementation was using the same RRC prototype filter on
transmit and receive filterbanks. But, the receive input from the device
is 6.25 Msps vs 3.84 Msps, which leads to wider RRC filter bandwidth
than specified and excess noise being present in the signal. Correct by
scaling the time domain pulse by a factor of 384/625 to obtain the
appropriate pulse shape.

Signed-off-by: Tom Tsou <tom@tsou.cc>